requirements.txt
# ============================================================
# AI音乐生成项目 - 依赖配置（方案A重构版）
# ============================================================

# # 核心依赖
# torch>=2.0.0
# torchaudio>=2.0.0

# 音频处理
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
noisereduce==3.0.0
audioread==3.0.1

# AI模型基础
# 说明：transformers 需要与模型 remote_code 相匹配
transformers>=4.45.0
accelerate>=0.25.0
huggingface_hub>=0.20.0
gradio_client>=1.3.0

# 数值计算
numpy==1.24.3
# py38 兼容：scipy 1.11.x 起通常不再提供 py38 轮子
scipy==1.10.1

# Web框架
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# 异步任务
celery==5.3.4
redis==5.0.1

# 数据库
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# 工具
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1

# 音频处理增强
praat-parselmouth==0.4.3
pyworld==0.3.4

# ============================================================
# 方案A架构新增依赖
# ============================================================

# Demucs（人声分离）
# 安装方式: pip install demucs
demucs>=4.0.0

# 音频处理增强
pedalboard>=0.7.7
pyloudnorm>=0.1.1

# 音频重采样
resampy>=0.4.2

# 音频可视化
matplotlib>=3.7.0

# ============================================================
# RVC / So-VITS-SVC 依赖
# ============================================================

# PyTorch Audio（用于RVC）
torchaudio>=2.0.0

# Fairseq（用于RVC的HuBERT模型）
fairseq>=0.10.2

# 音频编解码器
av>=11.0.0

# 音频特征提取
pyannote.audio>=3.1.0

# ============================================================
# 可选依赖（根据需要安装）
# ============================================================

# 高级音频效果
# pyrubberband>=0.3.0

# 实时音频处理
# pyaudio>=0.2.13

# GPU加速（需要CUDA）
# nvidia-cuda-nvrtc-cu11>=11.8.0
# nvidia-cudnn-cu11>=8.9.0

# ============================================================
# 模型相关（需要从源码安装）
# ============================================================

# ACE-Step（音乐生成）
# git+https://github.com/acestep/ace-step

# Fish Speech（TTS - 已弃用，保留用于兼容性）
# git+https://github.com/fishaudio/fish-speech

# RVC（音色转换）
# git+https://github.com/RVC-Project/Retrieval-based-Voice-Conversion-WebUI

# So-VITS-SVC（音色转换）
# git+https://github.com/svc-develop-team/so-vits-svc

# ============================================================
# 开发工具（可选）
# ============================================================

# 代码质量
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0

# 测试
# pytest>=7.0.0
# pytest-cov>=4.0.0

# 文档
# sphinx>=6.0.0
# sphinx-rtd-theme>=1.2.0
